% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baby_names}
\alias{baby_names}
\title{Yearly baby name frequencies for Ontario, Canada, 1917-2019.}
\format{
A data frame with 164,567 rows and 5 variables:
\describe{
  \item{year}{year}
  \item{name}{name}
  \item{freq}{frequency}
  \item{gender}{gender assigned at birth}
  \item{select_label}{concatenation of name and gender for package use}
  ...
}
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-female/}
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-male}
}
\usage{
baby_names
}
\description{
A dataset containing yearly counts of baby names in Ontario, Canada.
}
\keyword{datasets}
