% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baby_names_forselect}
\alias{baby_names_forselect}
\title{Baby names and genders, for internal package use}
\format{
An object of class \code{character} of length 8934.
}
\usage{
baby_names_forselect
}
\description{
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-female/}
}
\keyword{datasets}
