% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baby_names_top_all}
\alias{baby_names_top_all}
\title{Top 25 male and female baby names, 1917-2019}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5150 rows and 4 columns.
}
\usage{
baby_names_top_all
}
\description{
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-female/}
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-male}
}
\keyword{datasets}
