% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_yearly_names}
\alias{top_yearly_names}
\title{Top male and female baby names, 1917-2019}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 206 rows and 4 columns.
}
\usage{
top_yearly_names
}
\description{
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-female/}
* \url{https://data.ontario.ca/dataset/ontario-top-baby-names-male}
}
\keyword{datasets}
